entire codebase refactored, all functions re/moved from config.py, all tests pass
