+++ pyrate/core/config.py
@@ -23,9 +23,6 @@ in configuration files.
 # pylint: disable=W1203
 # pylint: disable=too-many-locals
 # pylint: disable=trailing-whitespace
-from typing import Dict
-import os
-from pyrate.core.logger import pyratelogger as _logger
 
 # general constants
 NO_MULTILOOKING = 1
@@ -232,86 +229,6 @@ TMPDIR = 'tmpdir'
 # Lookup to help convert args to correct type/defaults
 # format is	key : (conversion, default value)
 # None = no conversion
-PARAM_CONVERSION = {
-    #    REPROJECTION : (int, 3), # Default no conversion, CONVERSION NOT IMPLEMENTED
-    IFG_CROP_OPT: (int, 1),  # default to area 'intersection' option
-    IFG_LKSX: (int, NO_MULTILOOKING),
-    IFG_LKSY: (int, NO_MULTILOOKING),
-    IFG_XFIRST: (float, None),
-    IFG_XLAST: (float, None),
-    IFG_YFIRST: (float, None),
-    IFG_YLAST: (float, None),
-    NO_DATA_VALUE: (float, 0.0),
-
-    COH_MASK: (int, 0),
-    COH_THRESH: (float, 0.1),
-
-    REFX: (float, -1),
-    REFY: (float, -1),
-    REFNX: (int, 10),
-    REFNY: (int, 10),
-    REF_CHIP_SIZE: (int, 21),
-    REF_MIN_FRAC: (float, 0.5),
-    REF_EST_METHOD: (int, 1),  # default to average of whole image
-
-    ORBITAL_FIT: (int, 0),
-    ORBITAL_FIT_METHOD: (int, NETWORK_METHOD),
-    ORBITAL_FIT_DEGREE: (int, PLANAR),
-    ORBITAL_FIT_LOOKS_X: (int, 10),
-    ORBITAL_FIT_LOOKS_Y: (int, 10),
-
-    LR_NSIG: (int, 2),
-    # pixel thresh based on nepochs? not every project may have 20 epochs
-    LR_PTHRESH: (int, 3),
-    LR_MAXSIG: (int, 10),
-
-    # ATM_FIT: (int, 0), NOT CURRENTLY USED
-    # ATM_FIT_METHOD: (int, 2),
-
-    APSEST: (int, 0),
-    TLPF_CUTOFF: (int, 12),
-    TLPF_PTHR: (int, 1),
-
-    SLPF_CUTOFF: (float, 1.0),
-    SLPF_NANFILL: (int, 0),
-
-    DEMERROR: (int, 0),
-    # pixel thresh based on nepochs
-    DE_PTHR: (int, 3),
-
-    # pixel thresh based on nepochs? not every project may have 20 epochs
-    TIME_SERIES_PTHRESH: (int, 3),
-    TIME_SERIES_SM_FACTOR: (float, -1.0),
-    TIME_SERIES_SM_ORDER: (int, None),
-    TIME_SERIES_METHOD: (int, 2),  # Default to SVD method
-
-    PARALLEL: (int, 0),
-    PROCESSES: (int, 8),
-    PROCESSOR: (int, None),
-    NAN_CONVERSION: (int, 0),
-    NO_DATA_AVERAGING_THRESHOLD: (float, 0.0),
-}
-
-PATHS = [
-    OBS_DIR,
-    IFG_FILE_LIST,
-    DEM_FILE,
-    DEM_HEADER_FILE,
-    OUT_DIR,
-    SLC_DIR,
-    HDR_FILE_LIST,
-    COH_FILE_DIR,
-    COH_FILE_LIST,
-    BASE_FILE_DIR,
-    BASE_FILE_LIST,
-    LT_FILE,
-    APS_INCIDENCE_MAP,
-    APS_ELEVATION_MAP,
-]
-
-DEFAULT_TO_OBS_DIR = [SLC_DIR, COH_FILE_DIR, BASE_FILE_DIR]
-
-INT_KEYS = [APS_CORRECTION, APS_METHOD]
 
 # filenames reused in  many parts of the program
 REF_PIXEL_FILE = 'ref_pixel_file'
@@ -321,383 +238,3 @@ APS_ERROR_DIR = 'aps_error'
 PHASE_CLOSURE_DIR = 'phase_closure_dir'
 MST_DIR = 'mst_dir'
 TEMP_MLOOKED_DIR = 'temp_mlooked_dir'
-
-
-def get_config_params(path: str) -> Dict:
-    """
-    Reads the parameters file provided by the user and converts it into
-    a dictionary.
-
-    Args:
-        path: Absolute path to the parameters file.
-        validate: Validate the parameters if True, otherwise skip validation.
-        step: The current step of the PyRate workflow.
-
-    Returns:
-       A dictionary of parameters.
-    """
-    txt = ''
-    with open(path, 'r') as inputFile:
-        for line in inputFile:
-            if any(x in line for x in PATHS):
-                pos = line.find('~')
-                if pos != -1:
-                    # create expanded line
-                    line = line[:pos] + os.environ['HOME'] + line[(pos + 1):]
-            txt += line
-    params = _parse_conf_file(txt)
-    params[TMPDIR] = os.path.join(os.path.abspath(params[OUT_DIR]), 'tmpdir')
-
-    return params
-
-
-def _parse_conf_file(content) -> Dict:
-    """
-    Converts the parameters from their text form into a dictionary.
-
-    Args:
-        content: Parameters as text.
-
-    Returns:
-        A dictionary of parameters.
-    """
-
-    def _is_valid(line):
-        """
-        Check if line is not empty or has % or #
-        """
-        return line != "" and line[0] not in "%#"
-
-    lines = [ln.split() for ln in content.split('\n') if _is_valid(ln)]
-
-    # convert "field:   value" lines to [field, value]
-    kvpair = [(e[0].rstrip(":"), e[1]) for e in lines if len(e) == 2] + \
-             [(e[0].rstrip(":"), None) for e in lines if len(e) == 1]
-    parameters = dict(kvpair)
-    for p in PATHS:
-        if p not in parameters:
-            parameters[p] = None
-
-    for p in INT_KEYS:
-        if p not in parameters:
-            parameters[p] = '0'  # insert dummies
-
-    parameters = _handle_extra_parameters(parameters)
-
-    if not parameters:
-        raise ConfigException('Cannot parse any parameters from config file')
-
-    return _parse_pars(parameters)
-
-
-def _handle_extra_parameters(params):
-    """
-    Function to check if requirements for weather model correction are given.
-    """
-    params[APS_INCIDENCE_EXT] = None
-    params[APS_ELEVATION_EXT] = None
-
-    if params[APS_INCIDENCE_MAP] is not None:
-        params[APS_INCIDENCE_EXT] = \
-            os.path.basename(params[APS_INCIDENCE_MAP]).split('.')[-1]
-        params[APS_ELEVATION_MAP] = None
-        params[APS_ELEVATION_EXT] = None
-        return params
-
-    # define APS_ELEVATON_EXT for gamma prepifg
-    if params[APS_ELEVATION_MAP] is not None:
-        params[APS_ELEVATION_EXT] = os.path.basename(
-            params[APS_ELEVATION_MAP]).split('.')[-1]
-
-    return params
-
-
-def _parse_pars(pars) -> Dict:
-    """
-    Takes dictionary of parameters, converting values to required type
-    and providing defaults for missing values.
-
-    Args:
-        pars: Parameters dictionary.
-
-    Returns:
-        Dictionary of converted (and optionally validated) parameters.
-    """
-    # Fallback to default for missing values and perform conversion.
-    for k in PARAM_CONVERSION:
-        if pars.get(k) is None:
-            pars[k] = PARAM_CONVERSION[k][1]
-            # _logger.warning(f"No value found for parameter '{k}'. Using "f"default value {pars[k]}.")
-        else:
-            conversion_func = PARAM_CONVERSION[k][0]
-            if conversion_func:
-                try:
-                    pars[k] = conversion_func(pars[k])
-                except ValueError as e:
-                    _logger.error(
-                        f"Unable to convert '{k}': {pars[k]} to " f"expected type {conversion_func.__name__}.")
-                    raise e
-
-    # Fallback to default for missing paths.
-    for p in DEFAULT_TO_OBS_DIR:
-        if pars.get(p) is None:
-            pars[p] = pars[OBS_DIR]
-
-    return pars
-
-
-# ==== PARAMETER VALIDATION ==== #
-
-_PARAM_VALIDATION = {
-    IFG_FILE_LIST: (
-        lambda a: a is not None and os.path.exists(a),
-        f"'{IFG_FILE_LIST}': file must be provided and must exist."
-    ),
-    DEM_FILE: (
-        lambda a: a is not None and os.path.exists(a),
-        f"'{DEM_FILE}': file must be provided and must exist."
-    ),
-    DEM_HEADER_FILE: (
-        lambda a: a is not None and os.path.exists(a),
-        f"'{DEM_HEADER_FILE}': file must be provided and must exist."
-    ),
-    OUT_DIR: (
-        lambda a: a is not None,
-        f"'{OBS_DIR}': directory must be provided."
-    ),
-    APS_INCIDENCE_MAP: (
-        lambda a: os.path.exists(a) if a is not None else True,
-        f"'{APS_INCIDENCE_MAP}': file must exist."
-    ),
-    APS_ELEVATION_MAP: (
-        lambda a: os.path.exists(a) if a is not None else True,
-        f"'{APS_ELEVATION_MAP}': file must exists."
-    ),
-    IFG_CROP_OPT: (
-        lambda a: a in (1, 2, 3, 4),
-        f"'{IFG_CROP_OPT}': must select option 1, 2, 3, or 4."
-    ),
-    IFG_LKSX: (
-        lambda a: a >= 1,
-        f"'{IFG_LKSX}': must be >= 1."
-    ),
-    IFG_LKSY: (
-        lambda a: a >= 1,
-        f"'{IFG_LKSY}': must be >= 1."
-    ),
-    NO_DATA_VALUE: (
-        lambda a: True,
-        "Any float value valid."
-    ),
-    COH_MASK: (
-        lambda a: a in (0, 1),
-        f"'{COH_MASK}': must select option 0 or 1."
-    ),
-    REFX: (
-        lambda a: True,
-        "Any float value valid."
-    ),
-    REFY: (
-        lambda a: True,
-        "Any float value valid."
-    ),
-    ORBITAL_FIT: (
-        lambda a: a in (0, 1),
-        f"'{ORBITAL_FIT}': must select option 0 or 1."
-    ),
-    LR_NSIG: (
-        lambda a: 1 <= a <= 10,
-        f"'{LR_NSIG}': must be between 1 and 10 (inclusive)."
-    ),
-    LR_PTHRESH: (
-        lambda a: a >= 1,
-        f"'{LR_PTHRESH}': must be >= 1"
-    ),
-    LR_MAXSIG: (
-        lambda a: 0 <= a <= 1000,
-        f"'{LR_MAXSIG}': must be between 0 and 1000 (inclusive)."
-    ),
-    APSEST: (
-        lambda a: a in (0, 1),
-        f"'{APSEST}': must select option 0 or 1."
-    ),
-    DEMERROR: (
-        lambda a: a in (0, 1),
-        f"'{DEMERROR}': must select option 0 or 1."
-    ),
-    PARALLEL: (
-        lambda a: a in (0, 1),
-        f"'{PARALLEL}': must select option 0 or 1."
-    ),
-    PROCESSES: (
-        lambda a: a >= 1,
-        f"'{PROCESSES}': must be >= 1."
-    ),
-    PROCESSOR: (
-        lambda a: a in (0, 1, 2),
-        f"'{PROCESSOR}': must select option 0 or 1."
-    ),
-    NAN_CONVERSION: (
-        lambda a: a in (0, 1),
-        f"'{NAN_CONVERSION}': must select option 0 or 1."
-    ),
-    NO_DATA_AVERAGING_THRESHOLD: (
-        lambda a: True,
-        "Any float value valid."),
-}
-"""dict: basic validation functions for compulsory parameters."""
-
-_CUSTOM_CROP_VALIDATION = {
-    IFG_XFIRST: (
-        lambda a: a is not None,
-        f"'{IFG_XFIRST}': must be provided."
-    ),
-    IFG_XLAST: (
-        lambda a: a is not None,
-        f"'{IFG_XLAST}': must be provided."
-    ),
-    IFG_YFIRST: (
-        lambda a: a is not None,
-        f"'{IFG_YFIRST}': must be provided."
-    ),
-    IFG_YLAST: (
-        lambda a: a is not None,
-        f"'{IFG_YLAST}': must be provided.."
-    ),
-}
-"""dict: basic validation functions for custom cropping parameters."""
-
-_GAMMA_VALIDATION = {
-    HDR_FILE_LIST: (
-        lambda a: a is not None and os.path.exists(a),
-        f"'{HDR_FILE_LIST}': file must be provided and must exist."
-    ),
-}
-"""dict: basic validation functions for gamma parameters."""
-
-_COHERENCE_VALIDATION = {
-    COH_THRESH: (
-        lambda a: 0.0 <= a <= 1.0,
-        f"'{COH_THRESH}': must be between 0.0 and 1.0 (inclusive)."
-    ),
-    COH_FILE_LIST: (
-        lambda a: a is not None and not os.path.exists(a),
-        f"'{COH_FILE_LIST}': if file is provided it must exist."
-    ),
-}
-"""dict: basic validation functions for coherence parameters."""
-
-_BASELINE_VALIDATION = {
-    BASE_FILE_LIST: (
-        lambda a: a is not None and not os.path.exists(a),
-        f"'{BASE_FILE_LIST}': if file is provided it must exist."
-    ),
-}
-"""dict: basic validation functions for baseline parameters."""
-
-_LOOKUPTABLE_VALIDATION = {
-    LT_FILE: (
-        lambda a: a is not None and not os.path.exists(a),
-        f"'{LT_FILE}': if file is provided it must exist."
-    ),
-}
-"""dict: basic validation functions for lookup table file."""
-
-_ORBITAL_FIT_VALIDATION = {
-    ORBITAL_FIT_METHOD: (
-        lambda a: a in (1, 2),
-        f"'{ORBITAL_FIT_METHOD}': must select option 1 or 2."
-    ),
-    ORBITAL_FIT_DEGREE: (
-        lambda a: a in (1, 2, 3),
-        f"'{ORBITAL_FIT_DEGREE}': must select option 1, 2 or 3."
-    ),
-    ORBITAL_FIT_LOOKS_X: (
-        lambda a: a >= 1,
-        f"'{ORBITAL_FIT_LOOKS_X}': must be >= 1."
-    ),
-    ORBITAL_FIT_LOOKS_Y: (
-        lambda a: a >= 1,
-        f"'{ORBITAL_FIT_LOOKS_Y}': must be >= 1."
-    ),
-}
-"""dict: basic validation fucntions for orbital error correction parameters."""
-
-_APSEST_VALIDATION = {
-    TLPF_CUTOFF: (
-        lambda a: a >= 1,
-        f"'{TLPF_CUTOFF}': must be >= 1."
-    ),
-    TLPF_PTHR: (
-        lambda a: a >= 1,
-        f"'{TLPF_PTHR}': must be >= 1."
-    ),
-    SLPF_CUTOFF: (
-        lambda a: a >= 0.001,
-        f"'{SLPF_CUTOFF}': must be >= 0.001."
-    ),
-    SLPF_NANFILL: (
-        lambda a: a in (0, 1),
-        f"'{SLPF_NANFILL}': must select option 0 or 1."
-    ),
-}
-"""dict: basic validation functions for atmospheric correction parameters."""
-
-DEMERROR_VALIDATION = {
-    DE_PTHR: (
-        lambda a: a >= 1,
-        f"'{TLPF_PTHR}': must be >= 1."
-    ),
-}
-"""dict: basic validation functions for DEM error correction parameters."""
-
-
-_TIME_SERIES_VALIDATION = {
-    TIME_SERIES_PTHRESH: (
-        lambda a: a >= 1,
-        f"'{TIME_SERIES_PTHRESH}': must be >= 1."
-    ),
-    TIME_SERIES_SM_FACTOR: (
-        lambda a: -5.0 <= a <= 0,
-        f"'{TIME_SERIES_SM_FACTOR}': must be between -5.0 and 0."
-    ),
-    TIME_SERIES_SM_ORDER: (
-        lambda a: a in (1, 2),
-        f"'{TIME_SERIES_SM_ORDER}': must select option 1 or 2."
-    ),
-    TIME_SERIES_METHOD: (
-        lambda a: a in (1, 2),
-        f"'{TIME_SERIES_METHOD}': must select option 1 or 2."
-    ),
-}
-"""dict: basic vaidation functions for time series parameters."""
-
-_REFERENCE_PIXEL_VALIDATION = {
-    REFNX: (
-        lambda a: 1 <= a <= 50,
-        f"'{REFNX}': must be between 1 and 50 (inclusive)."
-    ),
-    REFNY: (
-        lambda a: 1 <= a <= 50,
-        f"'{REFNY}': must be between 1 and 50 (inclusive)."
-    ),
-    REF_CHIP_SIZE: (
-        lambda a: 1 <= a <= 101 and a % 2 == 1,
-        f"'{REF_CHIP_SIZE}': must be between 1 and 101 (inclusive) and be odd."
-    ),
-    REF_MIN_FRAC: (
-        lambda a: 0.0 <= a <= 1.0,
-        f"'{REF_MIN_FRAC}': must be between 0.0 and 1.0 "
-        "(inclusive)."
-    ),
-    REF_EST_METHOD: (
-        lambda a: a in (1, 2),
-        f"'{REF_EST_METHOD}': must select option 1 or 2."
-    ),
-}
-
-
-class ConfigException(Exception):
-    """
-    Default exception class for configuration errors.
-    """
